Many little fixes to prevent minor warnings.

+++ include/emp/datastructs/hash_utils.hpp
@@ -106,7 +106,7 @@ namespace emp {
   // helper functions for murmur hash
   #ifndef DOXYGEN_SHOULD_SKIP_THIS
   namespace internal {
-    constexpr inline uint64_t rotate(const size_t x, const size_t r) noexcept {
+    constexpr inline uint64_t rotate(const uint64_t x, const uint64_t r) noexcept {
       return (x << r) | (x >> (64 - r));
     }
     constexpr inline void fmix64(uint64_t& k) noexcept {
@@ -127,13 +127,13 @@ namespace emp {
   /// @param key Span of bytes to hash.
   /// @param seed Optional seed.
   /// @return Hash of key.
-  constexpr inline size_t murmur_hash(
+  constexpr inline uint64_t murmur_hash(
     const std::span<const std::byte> key,
-    const size_t seed = 0
+    const uint64_t seed = 0
   ) noexcept {
     // define constants
-    const size_t numbytes = key.size();
-    const size_t nblocks = numbytes / 16;
+    const uint64_t numbytes = key.size();
+    const uint64_t nblocks = numbytes / 16;
     const uint64_t c1 = 0x87c37b91114253d5LLU;
     const uint64_t c2 = 0x4cf5ad432745937fLLU;
